+++ client/src/components/ThumbnailCard.js
@@ -56,6 +56,7 @@ export const ThumbnailCard = ({ product }) => {
 						<CardMedia
 							component="img"
 							width="100%"
+							height="100%"
 							image={image}
 							alt={name}
 							sx={{ cursor: 'pointer' }}
